feat: split dataArrayOfObject into smaller files
This commit refactors the dataArrayOfObject file by splitting its fixed
arrayOfData into smaller, more manageable pieces for improved maintainability.
The update focuses solely on dividing the files, and no major functionality
changes or new features have been introduced.
