+++ components/layouts/layoutApp/layout/layoutFooter/footerSidebar/footerSidebarMenu.tsx
@@ -1,6 +1,5 @@
 import { PrefetchRouterButton } from '@buttons/button/prefetchRouterButton';
 import { SvgIcon } from '@components/icons/svgIcon';
-import { DATA_SIDEBAR_MENU } from '@data/dataArrayOfObjects';
 import { BREAKPOINT } from '@data/dataTypesConst';
 import { STYLE_HOVER_SLATE_LIGHT } from '@data/stylePreset';
 import { useSidebarOpen } from '@states/layouts/hooks';
@@ -10,6 +9,7 @@ import { useRouter } from 'next/router';
 import { Fragment as FooterSidebarMenuFragment, Fragment as TotalNumberTodos } from 'react';
 import { useRecoilValue } from 'recoil';
 import { TodosCount } from './todosCount';
+import { DATA_SIDEBAR_MENU } from '@data/dataArrayOfObjects/dataSidebarMenu';
 
 export const FooterSidebarMenu = () => {
   const router = useRouter();
